+++ src/xrCore/rt_compressor.cpp
@@ -31,7 +31,7 @@ u32 rtc_compress(void* dst, u32 dst_len, const void* src, u32 src_len)
 		(lzo_byte*)dst, (lzo_uintp)&out_size,
 		rtc_wrkmem);
 	VERIFY(r == LZO_E_OK);
-	return out_size;
+	return (u32)out_size;
 }
 
 u32 rtc_decompress(void* dst, u32 dst_len, const void* src, u32 src_len)
@@ -42,5 +42,5 @@ u32 rtc_decompress(void* dst, u32 dst_len, const void* src, u32 src_len)
 		(lzo_byte*)dst, (lzo_uintp)&out_size,
 		rtc_wrkmem);
 	VERIFY(r == LZO_E_OK);
-	return out_size;
+	return (u32)out_size;
 }
